Fix #121 - use correct type name for floats in generated code (#126)

+++ python/rpdk/go/resolver.py
@@ -4,7 +4,7 @@ PRIMITIVE_TYPES = {
     "string": "string",
     "integer": "int",
     "boolean": "bool",
-    "number": "float",
+    "number": "float64",
     UNDEFINED: "interface{}",
 }
